Tests - Separated testing for Staff and Unit and added test code.
