--- tests/TestUnit.cs
@@ -1,14 +1,0 @@
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace Class_Scheduler.Tests.ParseTests
-{
-    [TestClass]
-    public class TestUnit
-    {
-        [TestMethod]
-        public void TestUnitOutput()
-        {
-
-        }
-    }
-}
